Begins refactor of serverside code
